Update patch set 3
Patch Set 3:

Added one more correction in the list of repos.. 

Since we migrated from 1 LDAP per repo, the list of repos need to only show the repo in question. I think is redundant to have this information at all but I think it was added for other LF projects.

Patch-set: 3
Label: Code-Review=0