Move apex policy editor to policy-gui
This review moves the code for the apex policy editor from the apex repo
to the GUI repo.

The docker part of the build will be moved in the next review because it
is intertwined with the other two legacy apex GUIs and needs to be
disentangled.

Subsequent reviews on the apex repo will remove this code from there.

Issue-ID: POLICY-2621
Change-Id: Ib9a43f463b839201656208a48ab48869a19f963a
Signed-off-by: liamfallon <liam.fallon@est.tech>