Upload policy feature
Implements an upload feature that can be enable/disabled through
configuration properties. An upload endpoint must be provided also by
configuration.
The upload has similar feature as the download, but it sends the file
to the provided endpoint instead, converting it to TOSCA YAML.
Introduces configuration properties with jersey-mp-config/
helidon-microprofile-config.
Creates an configuration endpoint so the front-end can read it.
Introduces a pub/sub feature in the front-end so it can asynchronous
read configuration information and send configuration ready message to
components that depends on it.

Issue-ID: POLICY-2751
Signed-off-by: andre.schmid <andre.schmid@est.tech>
Change-Id: I5e8e91616a7d4dc69e6739604fcd36528cce0028