Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Configuration properties we have been using policy/common code for this, this review introduces another way of doing configuration??
> 
> Is this only for apex native policies? What about our other policy types, both native and non-native?

About the configuration, yes, it is introducing another way to do it because I didn't know there was already a way to do it. Please, could you show me some example I could follow?

This is only for the APEX Editor GUI. It is allowing to upload a policy file created in that interface to a configured endpoint. It is reusing part of the Download feature already present, that gives the user a JSON file, and uses the same idea of the apex-pdp/auth/cli-editor to parse the JSON to TOSCA before uploading it.

Patch-set: 1