Update patch set 2
Patch Set 2:

(26 comments)

Hi all, I've tried to reply all the comments, let me know if I miss one.
About the Javascript tests, this project lacks it. We should track this in a new ticket, as we need to define a framework, configure the build and achieve the desired coverage.
Thank you all for the constructive comments.

Patch-set: 2