Update patch set 3
Patch Set 3:

(5 comments)

Hi all, I propose to leave the changes to reuse the ApexEditorCli code to another commit. I was taking a look there and it needs some refactoring to be reused and also does not have as much validation as this code for the Apex Config and Tosca Template files (I'm planning to move this code there too).

If you allow it, I will probably upstream the refactor in two weeks window. Thanks in advance.

Patch-set: 3