Update patch set 4
Patch Set 4:

sorry for the delayed comment, but we still have the dist folder inside the jars, which I believe it is not necessary. Also the js folder contains many files used in development, but not necessary after the build, as they are part of the bundle.js now. I believe the only js files necessary are related to those includes you have in the index, which are separate libraries:

<script src="../js/jquery/jquery-3.4.1.min.js"></script>
<script src="../js/jquery-ui-1.12.1/jquery-ui.js"></script>
<!-- fileMenu JS file -->
<script type="text/javascript" src="../js/fileMenu.js"></script>
<!-- showhideElement JS file -->
<script type="text/javascript" src="../js/showhideElement.js"></script>
<!--  edit area JS file -->
<script language="javascript" type="text/javascript" src="../js/edit_area/edit_area_full.js"></script>
<!-- generated bundle.js -->
<script type="text/javascript" src="bundle.js"></script>

I think we could track this clean up in another ticket.

Patch-set: 4
Reviewer: Gerrit User 5565 <5565@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>