Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> the sonar builds are failing now:
> 
> 09:06:28 [ERROR] The version of node.js (8) you have used to run this analysis is deprecated and we stopped accepting it.
> 09:06:28 [ERROR] Please update to at least node.js 10. You can find more information here: https://sonarcloud.io/documentation/upcoming/

I looked into why the sonar jobs are failing in policy-gui. The error we are seeing is that we need to upgrade Node JS from v8 to a minimum of v10. The version of Node we use in policy-gui is actually v12+ - I upgraded this sometime in November of 2020 - I believe.
If you see here, it says: "The version of Node.js installed in the scanner environment should be upgraded to at least Node.js 10 by 1 February 2021." "The change applies specifically to the version of Node.js installed and used in the context where your SonarCloud scanner analysis tool is run, either in your local build environment or in your CI service. It does not have any impact on the Node.js version targeted by your project code."
Reading further into it, you can still use Node JS v8 or less in your code and it will not effect the sonar scanning. The issue is stemming from the version of Node JS installed/used by our analysis tool or Jenkins.
https://sonarcloud.io/documentation/appendices/end-of-support/
https://sonarcloud.io/documentation/appendices/move-analysis-node-js-10/
This page above outlines some steps to take to solve this issue. It only effects projects that are analyzing JavaScript or TypeScript and this is why we don't see this issue arise in other sonar build jobs for other projects. I'm not sure who needs to perform this upgrade from Node JS v8 -> Node JS v10+.
From what I can understand, our version of Node in policy-gui has no bearing on this failing.

Patch-set: 2