Fix sonarqube code smells
Making changed regarding code smells. Some variables being declared twice while in scope.
renaming some modelFileName in ApexFiles.js that is being declared while there is a constant with the same name.
renamed some of the for loop control variables so that sonarqube doesnt see it as an issue
Made changes based on reviews

Issue-ID: POLICY-3095
Change-Id: If38934247a4e4ff9c48043feb3adde5505c46993
Signed-off-by: mkidd <michael.kidd@est.tech>