Update patch set 1
Patch Set 1:

Dependency bugs have been fixed.
There remains a bug in the url of the private repository to which JS is to be sent (most likely for use with npm install command). The address should be placed in an environment variable
npm.publish.url
Where can I find this url?

Patch-set: 1