Fix static web resource paths in gui-editor-apex
Many of the relative paths used in the HTML/CSS are incorrect in the build, e.g. index.html references "../js/fileMenu.js" instead of "js/fileMenu.js". When the project is built, maven will package the "js" and "css" directories in the same directory as index.html. This patch corrects the relative paths.

Issue-ID: POLICY-3601
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: Ifb9f4b428f3732cd601a3128df27d720a55f88fa