Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm not sure about commenting out....Let's see what the other committers think.

I think one of java processes could run in the foreground, and in that case I think there would be no need for ngnix.    But let me ask you, is there a review though in OOM subsystem for gui?  I was just looking and see no gui component chart.   If that's the case, *no* OOM work for GUI just usage in docker/docker-compose, I am ok with the way that you had it initially (as we wouldn't break requirements).   As a note, the GUI port should also be configurable at a later point at least when an OOM chart is done for it.

Patch-set: 3
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>