Add user id only when it was provided
During the policy upload call, the user id is being added as a query
parameter even if it was not provided, causing to provide the value
"null" as the userId parameter.
This change just add the userId to the call when it is available.

Issue-ID: POLICY-3796
Signed-off-by: andre.schmid <andre.schmid@est.tech>
Change-Id: I7bcc928f1ca144ee1ad05e7d1028ea08a15ebd42
(cherry picked from commit 08b6e0e63e4524d02507d6709710b53f87690cc0)