Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Thanks Daniel, can you comment the usage of "demo-clamp-keystore" and "demo-clamp-truststore".   It seems they make it into the jar.   In OOM, it is provided automatically to us at container initialization as well as the keys.    Just trying to understand if there is potential introduction of any security issues handling certificates.   I noticed there are some references to insecure handling of certificates.   It would be good to have Seb look at it too.

Demo keystore and truststore are used for local dev testing, so it runs out-of-the-box. My plan is to replace the docker image application.yml in OOM chart to set location of AAF key/truststore. Likewise, useInsecureTrustManager is needed for testing against local clamp backend using self-signed certs - this can also be fixed in OOM chart application.yml.
This is similar to the previous implementation with nginx where a test key was included in image, and the OOM chart replaced the nginx config file.
If you still think it's unwise to include these files in the JAR, I can create a maven profile that includes them for dev, but excludes for production/OOM build.

Patch-set: 4