Update patch set 1
Patch Set 1:

This review is part of POLICY-3896 - Apex Editor using Spring Boot.
It was necessary to do some restructuring of Apex Editor to facilitate migration to Spring Boot. Also, as part of POLICY-3638 (Run all GUIs in same jar), it will be necessary to break Apex Editor into separate frontend and backend modules. I am thus restructuring the frontend build here to reduce the complexity of further reviews.
As a bonus, the restructure simplifies the POM and fixes an issue where JS source and test files got copied to the JAR.

Patch-set: 1