Amend GUI to work directly towards ACM runtime
GUI proxies the URLs towards the ACM runtime server across CLAMP backend
because clamp-backend handles the certificates for HTTPS communication
and also handles AAF.

For now, the Policy GUI is being used for testing and demonstration
purposes only.

As AAF is disabled and we are using HTTP, the GUI can now work towards
the ACM runtime server port directly for test installations.

Issue-ID: POLICY-4122
Change-Id: Idca87d520a5eef2058781534df6f96a0260c5eaf
Signed-off-by: liamfallon <liam.fallon@est.tech>