Update Dockerfile for PF base image
The Policy Framework base image run under the "onap" user rather than
under the "root" user. This means that root access has to be enabled in
child Dockerfiles for configuraiton that requires root access.

Root access is turned off again later in the Dockerfile so that the
image will run under another user.

Issue-ID: POLICY-4558
Change-Id: I75e53c6646eb19cd09db7e658bb89e5a3d979c2e
Signed-off-by: liamfallon <liam.fallon@est.tech>